Use minipunch's version of vehicle inventories, essentialmode, and added usa-characters
This version of vehicle-inventories does not contain @distritic 's changes that allowed interaction with vehicle inventories of non-player owned vehicles. That will need to be re-added (preferably with a hash table with plate as the keys instead of an array).

Also added my latest version of essentialmode (to work with new characters system).

Also added usa-characters resource.