feat(Misc): Basketball (#1589)
* feat(Misc): Basketball

You can now go to certain places around the map which are marked on the map as "Basketball" in which you can play basketball with friends, or you can go to a hardware store and buy your own hoop to place down at a location of choice! (It's recommended that you pickup your basketball hoop when done using it so it doesn't get removed on server restarts)

There are 3 different basketball game modes which are HORSE, Around The World, and Free Throw. HORSE & Around The World have the option for wagers so you can wage other players, and free throw is just free throwing no money involved.

To leave an ongoing basketball game use `/leavebasketball`

* update hoop price

* Make blip smaller

* Reupload lol

* Update .gitignore

* Upload missing DB files

Co-authored-by: Josh Yolles <joshyolles@gmail.com>